refactor(core): pipeline-worker normalize-naver 핸들러 분리
615줄 거대 핸들러 분해 (3/4) — 네이버 댓글 병렬 수집(semaphore=4)을
별도 파일로 추출. since 맵 / 부분 실패 허용 / 배치 취소 확인 보존.

worker.test.ts: 단일 파일 grep을 pipeline-worker* 디렉토리 검색으로 갱신.

리팩토링 마스터플랜 Phase 2 PR 2-F.
